Simplify and fix module timeouts
This commit partially reverts 8f010cf51e9. Instead of going through an
ExecutorService to run modules, modules now run on the thread that
handles the HTTP request, which is a lot simpler. Timeouts are handled
by writing an own TimeoutInterrupter.

The fix here is that the old code never worked correctly. An error was
returned to the client when the timeout expired. However, the "module
worker thread" continued running which defeats the purpose of the
timeout.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>